Add dedicated text module for font and text rendering functionality
- Moved text and font functions from draw.c to new text.c file
- Added text_test.rb for text functionality testing
- Implemented missing mrb_cf_result_from_cf_result function
- Updated header files and initialization functions

This refactoring makes the codebase more modular and matches the project's organization pattern with dedicated files for different functionality components.
