+++ src/draw.c
@@ -1,7 +1,6 @@
 #include "draw.h"
 #include "aabb.h"
 #include "circle.h"
-#include "result.h"
 #include "sprite.h"
 #include "vector.h"
 
@@ -156,103 +155,6 @@ static mrb_value mrb_cf_draw_circle2(mrb_state* mrb, mrb_value self)
     return mrb_nil_value();
 }
 
-static mrb_value mrb_cf_draw_text(mrb_state* mrb, mrb_value self)
-{
-    mrb_value position_obj;
-    char* text;
-    mrb_int num_chars_to_draw = -1;
-
-    mrb_get_args(mrb, "zo|i", &text, &position_obj, &num_chars_to_draw);
-
-    CF_V2 position = *mrb_cf_v2_unwrap(mrb, position_obj);
-
-    cf_draw_text(text, position, (int)num_chars_to_draw);
-
-    return mrb_nil_value();
-}
-
-static mrb_value mrb_cf_make_font(mrb_state* mrb, mrb_value self)
-{
-    char* path;
-    char* font_name;
-
-    mrb_get_args(mrb, "zz", &path, &font_name);
-
-    CF_Result result = cf_make_font(path, font_name);
-
-    return mrb_cf_result_wrap_nested(mrb, &result);
-}
-
-static mrb_value mrb_cf_push_font(mrb_state* mrb, mrb_value self)
-{
-    char* font_name;
-
-    mrb_get_args(mrb, "z", &font_name);
-
-    cf_push_font(font_name);
-
-    return mrb_nil_value();
-}
-
-static mrb_value mrb_cf_pop_font(mrb_state* mrb, mrb_value self)
-{
-    cf_pop_font();
-    return mrb_nil_value();
-}
-
-static mrb_value mrb_cf_push_font_size(mrb_state* mrb, mrb_value self)
-{
-    mrb_float size;
-
-    mrb_get_args(mrb, "f", &size);
-
-    cf_push_font_size((float)size);
-
-    return mrb_nil_value();
-}
-
-static mrb_value mrb_cf_pop_font_size(mrb_state* mrb, mrb_value self)
-{
-    cf_pop_font_size();
-    return mrb_nil_value();
-}
-
-static mrb_value mrb_cf_text_width(mrb_state* mrb, mrb_value self)
-{
-    char* text;
-    mrb_int num_chars_to_draw = -1;
-
-    mrb_get_args(mrb, "z|i", &text, &num_chars_to_draw);
-
-    float width = cf_text_width(text, (int)num_chars_to_draw);
-
-    return mrb_float_value(mrb, width);
-}
-
-static mrb_value mrb_cf_text_height(mrb_state* mrb, mrb_value self)
-{
-    char* text;
-    mrb_int num_chars_to_draw = -1;
-
-    mrb_get_args(mrb, "z|i", &text, &num_chars_to_draw);
-
-    float height = cf_text_height(text, (int)num_chars_to_draw);
-
-    return mrb_float_value(mrb, height);
-}
-
-static mrb_value mrb_cf_text_size(mrb_state* mrb, mrb_value self)
-{
-    char* text;
-    mrb_int num_chars_to_draw = -1;
-
-    mrb_get_args(mrb, "z|i", &text, &num_chars_to_draw);
-
-    CF_V2 size = cf_text_size(text, (int)num_chars_to_draw);
-
-    return mrb_cf_v2_wrap_nested(mrb, &size);
-}
-
 void mrb_cute_draw_init(mrb_state* mrb, struct RClass* mCute)
 {
     mrb_define_module_function(mrb, mCute, "cf_draw_sprite", mrb_cf_draw_sprite, MRB_ARGS_REQ(1));
@@ -268,17 +170,4 @@ void mrb_cute_draw_init(mrb_state* mrb, struct RClass* mCute)
     mrb_define_module_function(mrb, mCute, "cf_draw_quad2", mrb_cf_draw_quad2, MRB_ARGS_REQ(6));
     mrb_define_module_function(mrb, mCute, "cf_draw_circle", mrb_cf_draw_circle, MRB_ARGS_REQ(2));
     mrb_define_module_function(mrb, mCute, "cf_draw_circle2", mrb_cf_draw_circle2, MRB_ARGS_REQ(3));
-    mrb_define_module_function(mrb, mCute, "cf_draw_text", mrb_cf_draw_text, MRB_ARGS_ARG(2, 1));
-
-    // Font functions
-    mrb_define_module_function(mrb, mCute, "cf_make_font", mrb_cf_make_font, MRB_ARGS_REQ(2));
-    mrb_define_module_function(mrb, mCute, "cf_push_font", mrb_cf_push_font, MRB_ARGS_REQ(1));
-    mrb_define_module_function(mrb, mCute, "cf_pop_font", mrb_cf_pop_font, MRB_ARGS_NONE());
-    mrb_define_module_function(mrb, mCute, "cf_push_font_size", mrb_cf_push_font_size, MRB_ARGS_REQ(1));
-    mrb_define_module_function(mrb, mCute, "cf_pop_font_size", mrb_cf_pop_font_size, MRB_ARGS_NONE());
-
-    // Text measurement functions
-    mrb_define_module_function(mrb, mCute, "cf_text_width", mrb_cf_text_width, MRB_ARGS_ARG(1, 1));
-    mrb_define_module_function(mrb, mCute, "cf_text_height", mrb_cf_text_height, MRB_ARGS_ARG(1, 1));
-    mrb_define_module_function(mrb, mCute, "cf_text_size", mrb_cf_text_size, MRB_ARGS_ARG(1, 1));
 }
