+++ src/text.h
@@ -1,0 +1,8 @@
+#ifndef MRB_CUTE_TEXT_H
+#define MRB_CUTE_TEXT_H
+
+#include "mrb_cute.h"
+
+void mrb_cute_text_init(mrb_state* mrb, struct RClass* mCute);
+
+#endif // MRB_CUTE_TEXT_H
